Transition the build system to CMake
The ./configure.sh script can be used to set up an out-of-tree build.

The envset.sh script replaces the make_environment_* scripts and should
be sourced by the user when using GRASP.

I only link BLAS/LAPACK directly against lib9290. In principle we should
link it against all libraries and applications that use them directly,
but this works as well.