cleanup: remove trailing whitespace and newlines
Modern editors generally remove these automatically, and hence it is
necessary to clean up all files. Otherwise future source change commits
will contain a lot of noise with with whitespace changes, or the
developer has to patch-add the source files, separating whitespace
changes from meaningful changes. As both are sub-optimal, the best
option is to fix all the whitespace at once.

The following commands were used to clean up the whitespace:

  find . -type f -exec sed --in-place -e :a -e '/^\n*$/{$d;N;};/\n$/ba' {} \;
  find . -type f -exec sed --in-place 's/[[:space:]]\+$//' {} \;

Note: when using these commands, make sure that you do not have a .git
directory in the search path -- it will corrupt the Git database!