Add support for CMakeLists.user
Additional programs that integrate with GRASP and use CMake, and also
other additional configuration, can be now be put dropped into the
CMakeLists.user file, which is automatically included in the main
CMakeLists.txt when present.

This means that the users can avoid having to modify the version
controlled main CMakeLists.txt.