Change license text to standard MIT
This is so that automated systems could recognize that this indeed is
the MIT license.

None of the rights or restrictions granted by the license were changed,
so the updated wording does not change the spirit of the license in any
way. It is clear that "associated documentation files" does not refer to
the separate manual, as the manual is not part of the repository and has
its own copyright and license statement.