Fix typos and add formatting for README/docs
Fix the project name in Doxygen's config.

Also, remove GRASP acronym from README title The title is already long
enough. Also, Doxygen uses that as the page title, together with the
project name (so without this the page title becomes "GRASP: GRASP:
General Relativistic ..."). With this patch, the duplication in the
title is removed.

Finally, fix typos and add some formatting to the README.